APEXCORE-627 : Unit test AtMostOnceTest intermittently fails
Fixed a race problem for calling checkpointed and committed in RecoverableInputOperator. The original implementation of the class used the method checkpointed() of the interface CheckpointListener to refresh a value of one of the criteria variables checkpointedWindowId. The fix update uses the method beforeCheckpoint() of the interface CheckpointNotificationListener. It guaranties that the update of the variable checkpointedWindowId will be done before the call of the method committed().