APEXCORE-704 Add supporting of programmatic logger appender
Implemented supporting of a programmatic logger appender that can be added to Apex Application Master and Containers and be configurable programmatically. The new programmatic appender can be defined in Java code or via a value of the new Apex attribute "LOGGER_APPENDER".

The syntax of the attribute value: {appender-names};{logger-properties}