APEXCORE-626 shutdown-app in cli supports app-name as argument
The command shutdown-app now supports not only the app-id but also the
app name as an argument.
In case of multiple appIds or appNames the shutdown-app command follows
a "best effort" approach, meaning that it tries to shutdown each app
individually and prints an error message if the shutdown of an app
fails, but continues to shutdown all the others.

This commit adds the described functionality, documentation to the cli
help messages and a test to ensure correct best effort behaviour.