APEXCORE-767 Set parent classloader in StramAppLauncher loadDependencies
StramAppLauncher.loadDependencies is called multiple times when starting
an application via the apex-cli with the -local option. In each of the
calls to loadDependencies, the contextClassLoader of the current thread
would be replaced with a new instance of URLClassLoader (which has no
parent class loader set).

This can lead to issues, e.g. when one aquires the current
contextClassLauncher, loads a class with it and tries to cast it to a
class which was loaded with a previous version of the contextClassLoader,
resulting in a ClassCastException.

An example of this bug can be seen in APEXMALHAR-2511

The changes in this commit fix this by passing the parent class loader
for each new instance of URLClassLoader to the current
contextClassLoader